Move session token persistence into session module
